Makes Tone component out of theremin instance
Making the theremin imperative interface declarative through a React
component. Wanting to add in some synced state, hooked up the pitch to
the number of clicks.

Using `Things.update` through the console is an easy way to bring the
pitch back down when it inevitably gets too high =)

Challenge: How can we hook up buttons to raise and lower the pitch?